requirements.txt
# 数式抽出パイプライン (extract_equations.py)
# インストール: pip install -r requirements.txt

google-genai>=1.0.0
pydantic>=2.0.0
requests>=2.28.0
torch>=2.0.0
transformers>=4.40.0
torch-geometric>=2.5.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
networkx>=3.0
